+++ packages/backends/hive-gateway/src/email/index.ts
@@ -1,46 +1,39 @@
-import { SES } from 'aws-sdk'
 import {compile} from 'handlebars'
 import fs, { readFileSync } from 'fs';
 import path from 'path';
 
-const ses = new SES({region: 'ap-southeast-2'})
+import nodemailer from "nodemailer";
 
-export const sendMail = (
+
+export const sendMail = async (
+	transporter: nodemailer.Transporter,
 	to: string[], 
 	subject: string, 
 	body: string,
 	htmlTemplate?: string,
 	htmlObject?: any
 ) => {
-	return new Promise((resolve, reject) => {
-		ses.sendEmail({
-			Source: "HexHive Systems <no-reply@hexhive.io>",
-			Destination: {
-				ToAddresses: to
-			},
-			Message: {
-				Subject: {Data: subject},
-				Body: {
-					Html: htmlTemplate ? {
-						Data: compile(htmlTemplate)(htmlObject) 
-					}: undefined, 
-					Text: {Data: body}
-				}
-			}
-		}, (err, data) => {
-			if(err) return reject(err)
-			resolve(data)
+
+		return await transporter.sendMail({
+			from: "HexHive Systems <no-reply@hexhive.io>",
+			to,
+			subject,
+			text: body,
+			html: htmlTemplate ? compile(htmlTemplate)(htmlObject) : undefined		
 		})
-	})
 
 }
 
-export const sendInvite = async (invite: {link: string, sender: string, to: string, receiver: string, type: string}) => {
+export const sendInvite = async (
+	transporter: nodemailer.Transporter,
+	invite: {link: string, sender: string, to: string, receiver: string, type: string}
+) => {
 	const inviteFile = invite.type == 'User' ? 'invite-user.html' : 'invite-collaborator.html';
 	const htmlTemplate = readFileSync(path.join(__dirname, './templates/', inviteFile), 'utf-8')
 
 	const subject = `${invite.sender} has invited you to an organisation on HexHive`
 	return await sendMail(
+		transporter,
 		[invite.to], 
 		subject,
 		`Kia Ora${invite.receiver ? ` ${invite.receiver}`: ''},
@@ -55,9 +48,13 @@ export const sendInvite = async (invite: {link: string, sender: string, to: stri
 		)
 }
 
-export const sendPasswordReset = async (reset_event: {requested: string, link: string}) => {
+export const sendPasswordReset = async (
+	transporter: nodemailer.Transporter,
+	reset_event: {requested: string, link: string}
+) => {
 	const subject = `Password reset request`
 	return await sendMail(
+		transporter,
 		[reset_event.requested], 
 		subject,
 		`Kia Ora,