+++ packages/backends/hive-gateway/src/schema/subschema/index.ts
@@ -1,12 +1,12 @@
 import { mergeResolvers } from '@graphql-tools/merge'
 import { PrismaClient } from '@hexhive/data'
-
+import nodemailer from 'nodemailer'
 import { JSONDefinition, JSONResolver } from 'graphql-scalars'
 import acl from './acl'
 import apps from './apps'
 
-export default (prisma: PrismaClient, schemas: { [key: string]: {acl: any[]} }) => {
-    const { typeDefs: aclTypeDefs, resolvers: aclResolvers } = acl(prisma)
+export default (prisma: PrismaClient, schemas: { [key: string]: {acl: any[]} }, transporter?: nodemailer.Transporter) => {
+    const { typeDefs: aclTypeDefs, resolvers: aclResolvers } = acl(prisma, transporter)
     const { typeDefs: appTypeDefs, resolvers: appResolvers }  = apps(prisma, schemas);
 
     const resolvers : any = mergeResolvers([