+++ packages/core/hexhive-data/src/user/index.ts
@@ -2,6 +2,7 @@
 // import { Session, Transaction } from "neo4j-driver";
 // import { createHash } from 'crypto';
 
+
 // export const createUser = async (tx: Transaction, user: {
 // 	name: string, 
 // 	username: string,
@@ -86,3 +87,4 @@
 
 // 	return result.records?.[0]?.get(0).properties
 // }
+