+++ packages/infrastructure/apps/src/integration-clients/hive-flow-mssql/index.ts
@@ -37,7 +37,7 @@ export const HiveFlowIntegration = (provider: k8s.Provider) => {
                             // { name: 'UI_URL',  value: `https://${domainName}/dashboard` },
                             // { name: 'BASE_URL',  value: `https://${domainName}`},
                             { name: 'ORG_ID', value: process.env.FLOW_ORG_ID },
-                            { name: "NEO4J_URI", value: process.env.NEO4J_URI /*neo4Url.apply((url) => `neo4j://${url}.default.svc.cluster.local`)*/ },
+                            // { name: "NEO4J_URI", value: process.env.NEO4J_URI /*neo4Url.apply((url) => `neo4j://${url}.default.svc.cluster.local`)*/ },
                             // { name: "MONGO_URL", value: mongoUrl.apply((url) => `mongodb://${url}.default.svc.cluster.local`) },
                         ]
                         // livenessProbe: {