+++ packages/backends/hive-gateway/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.2-beta.70](https://github.com/TheTechCompany/HexHive/compare/v1.0.2-beta.69...v1.0.2-beta.70) (2023-11-17)
+
+**Note:** Version bump only for package @hexhive/gateway
+
+## [1.0.2-beta.69](https://github.com/TheTechCompany/HexHive/compare/v1.0.2-beta.68...v1.0.2-beta.69) (2023-09-24)
+
+**Note:** Version bump only for package @hexhive/gateway
+
+## [1.0.2-beta.68](https://github.com/TheTechCompany/HexHive/compare/v1.0.2-beta.67...v1.0.2-beta.68) (2023-09-24)
+
+**Note:** Version bump only for package @hexhive/gateway
+
+## [1.0.2-beta.67](https://github.com/TheTechCompany/HexHive/compare/v1.0.2-beta.66...v1.0.2-beta.67) (2023-09-24)
+
+**Note:** Version bump only for package @hexhive/gateway
+
+## [1.0.2-beta.66](https://github.com/TheTechCompany/HexHive/compare/v1.0.2-beta.65...v1.0.2-beta.66) (2023-09-24)
+
+**Note:** Version bump only for package @hexhive/gateway
+
+## [1.0.2-beta.64](https://github.com/TheTechCompany/HexHive/compare/v1.0.2-beta.63...v1.0.2-beta.64) (2023-09-24)
+
+**Note:** Version bump only for package @hexhive/gateway
+
 ## [1.0.2-beta.63](https://github.com/TheTechCompany/HexHive/compare/v1.0.2-beta.62...v1.0.2-beta.63) (2023-07-24)
 
 **Note:** Version bump only for package @hexhive/gateway